# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BusinessNow PSA Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rate-cards:
    get:
      operationId: listRateCards
      tags: [ratecards]
      summary: List rate cards
      responses:
        "200":
          description: List of rate cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RateCard"
    post:
      operationId: createRateCard
      tags: [ratecards]
      summary: Create a rate card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRateCardBody"
      responses:
        "201":
          description: Created rate card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateCard"
  /rate-cards/{id}:
    patch:
      operationId: updateRateCard
      tags: [ratecards]
      summary: Update rate card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRateCardBody"
      responses:
        "200":
          description: Updated rate card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateCard"
    delete:
      operationId: deleteRateCard
      tags: [ratecards]
      summary: Delete rate card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── RESOURCES ────────────────────────────────────────────────
...
  /projects/{id}/portal-tokens:
    post:
      summary: Create or regenerate a client portal token
      operationId: createPortalToken
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePortalTokenBody'
      responses:
        '201':
          description: Token created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortalToken'
    get:
      summary: List portal tokens for a project
      operationId: listPortalTokens
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PortalToken'
  /projects/{id}/portal-tokens/{tokenId}:
    delete:
      summary: Revoke a portal token
      operationId: revokePortalToken
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: path
          name: tokenId
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Revoked
  /projects/{id}/gantt:
    get:
      summary: Get Gantt data for a project
      operationId: getProjectGantt
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GanttData'
  /portal/{token}:
    get:
      summary: Get public client portal data by token
      operationId: getPortal
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortalData'
        '404':
          description: Token not found or expired
  /prospects:
    get:
      summary: List prospects
      operationId: listProspects
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: ownerId
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Prospect'
    post:
      summary: Create prospect
      operationId: createProspect
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProspectBody'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prospect'
  /prospects/{id}:
    get:
      summary: Get prospect
      operationId: getProspect
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prospect'
    patch:
      summary: Update prospect
      operationId: updateProspect
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProspectBody'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prospect'
    delete:
      summary: Delete prospect
      operationId: deleteProspect
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
  /prospects/{id}/convert:
    post:
      summary: Convert prospect to customer account
      operationId: convertProspect
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConvertProspectBody'
      responses:
        '200':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConvertProspectResult'
...
components:
  schemas:
...
    Account:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        domain:
          type: string
        tier:
          type: string
        region:
          type: string
        status:
          type: string
        contractValue:
          type: number
        billingAddress:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        convertedFromProspectId:
          type: integer
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - domain
        - tier
        - region
        - status
        - contractValue
        - createdAt
        - updatedAt
...
    RateCard:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        currency:
          type: string
        status:
          type: string
        effectiveDate:
          type: string
          nullable: true
        defaultRate:
          type: number
          nullable: true
        roles:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
              rate:
                type: number
            required:
              - role
              - rate
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - currency
        - status
        - roles
        - createdAt
        - updatedAt
    CreateRateCardBody:
      type: object
      properties:
        name:
          type: string
        currency:
          type: string
        roles:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
              rate:
                type: number
            required:
              - role
              - rate
      required:
        - name
        - currency
        - roles
    UpdateRateCardBody:
      type: object
      properties:
        name:
          type: string
        currency:
          type: string
        roles:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
              rate:
                type: number
            required:
              - role
              - rate

    # ─── ALLOCATION ─────────────────────────────────────────────
...
    PortalToken:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        token:
          type: string
        label:
          type: string
          nullable: true
        isActive:
          type: boolean
        expiresAt:
          type: string
          nullable: true
        viewCount:
          type: integer
        lastViewedAt:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - projectId
        - token
        - isActive
        - viewCount
        - createdAt
        - updatedAt
    CreatePortalTokenBody:
      type: object
      properties:
        label:
          type: string
        expiresAt:
          type: string
          nullable: true
    GanttRow:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        name:
          type: string
        startDate:
          type: string
          nullable: true
        dueDate:
          type: string
          nullable: true
        status:
          type: string
        completion:
          type: integer
        parentId:
          type: integer
          nullable: true
        isMilestone:
          type: boolean
      required:
        - id
        - type
        - name
        - status
        - completion
        - isMilestone
    GanttData:
      type: object
      properties:
        projectId:
          type: integer
        projectStart:
          type: string
        projectEnd:
          type: string
        rows:
          type: array
          items:
            $ref: '#/components/schemas/GanttRow'
      required:
        - projectId
        - projectStart
        - projectEnd
        - rows
    PortalMilestone:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          type: string
        dueDate:
          type: string
          nullable: true
        completedAt:
          type: string
          nullable: true
      required:
        - id
        - name
        - status
    PortalDocument:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - type
        - updatedAt
    PortalData:
      type: object
      properties:
        project:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            status:
              type: string
            health:
              type: string
            completion:
              type: integer
            startDate:
              type: string
            dueDate:
              type: string
            description:
              type: string
              nullable: true
            accountName:
              type: string
              nullable: true
          required:
            - id
            - name
            - status
            - health
            - completion
            - startDate
            - dueDate
        milestones:
          type: array
          items:
            $ref: '#/components/schemas/PortalMilestone'
        documents:
          type: array
          items:
            $ref: '#/components/schemas/PortalDocument'
        overallProgress:
          type: number
        totalTasks:
          type: integer
        completedTasks:
          type: integer
        label:
          type: string
          nullable: true
      required:
        - project
        - milestones
        - documents
        - overallProgress
        - totalTasks
        - completedTasks
    Prospect:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        contactName:
          type: string
          nullable: true
        contactEmail:
          type: string
          nullable: true
        contactPhone:
          type: string
          nullable: true
        status:
          type: string
        source:
          type: string
          nullable: true
        estimatedValue:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
        convertedAccountId:
          type: integer
          nullable: true
        convertedAt:
          type: string
          nullable: true
        ownerId:
          type: integer
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - status
        - createdAt
        - updatedAt
    CreateProspectBody:
      type: object
      properties:
        name:
          type: string
        contactName:
          type: string
          nullable: true
        contactEmail:
          type: string
          nullable: true
        contactPhone:
          type: string
          nullable: true
        status:
          type: string
        source:
          type: string
          nullable: true
        estimatedValue:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
        ownerId:
          type: integer
          nullable: true
      required:
        - name
    UpdateProspectBody:
      type: object
      properties:
        name:
          type: string
        contactName:
          type: string
          nullable: true
        contactEmail:
          type: string
          nullable: true
        contactPhone:
          type: string
          nullable: true
        status:
          type: string
        source:
          type: string
          nullable: true
        estimatedValue:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
        ownerId:
          type: integer
          nullable: true
    ConvertProspectBody:
      type: object
      properties:
        tier:
          type: string
        region:
          type: string
        domain:
          type: string
        contractValue:
          type: number
      required:
        - tier
        - region
        - domain
    ConvertProspectResult:
      type: object
      properties:
        account:
          $ref: '#/components/schemas/Account'
        prospectId:
          type: integer
      required:
        - account
        - prospectId
...